Moved PinnableBufferCacheEventSource to shared
