+++ src/System.Private.CoreLib/src/System/Threading/Overlapped.cs
@@ -19,6 +19,7 @@
 **
 =============================================================================*/
 
+using System;
 using System.Runtime;
 using System.Runtime.InteropServices;
 using System.Runtime.CompilerServices;
